refactor(ui): extract settings/lifecycle.tsx — mount + open/close/save/init + content binders
settings.tsx is now a ~20-line barrel re-exporting the public surface from
settings/data/config-loader.ts and settings/lifecycle.tsx. All implementation
lives under settings/:

  settings/
    lifecycle.tsx                  — mount, open/close/save, init, SettingsDialogContent/ShellBound/ObserverStatusBanner binders, model accessor wrappers, event-handler factory instance
    components/                    — 11 leaf + composite components (one per file)
    data/                          — pure helpers, state object, state-ops, config-loader, form-state, collect-payload, diff-payload, event-handlers factory, formatters, parsers, types, dom helpers, model accessors, notice builder
    hooks/use-help-tooltip.tsx     — shared help-tooltip hook

Completes the settings.tsx split — the original 2167-LOC god file is fully
decomposed.
